+++ back/Subasta.core/services/DepartamentoService.cs
@@ -13,15 +13,24 @@ using System.Text;
 
 namespace Subasta.core.services
 {
-    public class DepartamentoService: IDepartamentoService
+    public class DepartamentoService : IDepartamentoService
     {
         readonly IMapper mapper;
         readonly IUnitOfWork uowService;
-
-        public DepartamentoService(IMapper mapper, IUnitOfWork uowService)
+        readonly IClienteRepository clienteRepository;
+        readonly IEventoRepository eventoRepository;
+        readonly ILoteRepository loteRepository;
+        readonly IAnimalRepository animalRepository;
+        public DepartamentoService(IMapper mapper, IUnitOfWork uowService,
+          IClienteRepository clienteRepository, IEventoRepository eventoRepository,
+            ILoteRepository loteRepository, IAnimalRepository animalRepository)
         {
             this.mapper = mapper;
             this.uowService = uowService;
+            this.clienteRepository = clienteRepository;
+            this.eventoRepository = eventoRepository;
+            this.loteRepository = loteRepository;
+            this.animalRepository = animalRepository;
         }
 
         public void Add(DepartamentoDto dto)
@@ -42,12 +51,42 @@ namespace Subasta.core.services
             }
         }
 
+        public int AddWithReturn(DepartamentoDto dto)
+        {
+            try
+            {
+                return uowService.DepartamentoRepository.Add(mapper.Map<Departamento>(dto));
+            }
+            catch (ExceptionData)
+            {
+                throw;
+            }
+            catch (Exception ex)
+            {
+
+                throw new ExceptionCore("error al intentar agregar el departamento", ex);
+            }
+        }
+
         public void Delete(DepartamentoDto entity)
         {
             try
             {
-                uowService.DepartamentoRepository.Delete(mapper.Map<Departamento>(entity));
-                uowService.Save();
+                var clientes = clienteRepository.GetAll().Where(c => c.Municipio.DepartamentoId == entity.DepartamentoId);
+                var eventos = eventoRepository.GetAll().Where(d => d.Municipio.DepartamentoId == entity.DepartamentoId);
+                var lotes = loteRepository.GetAll().Where(d => d.Municipio.DepartamentoId == entity.DepartamentoId);
+                var animales = animalRepository.GetAll().Where(d => d.Municipio.DepartamentoId == entity.DepartamentoId);
+
+                if (clientes.Count() == 0 && eventos.Count() == 0 && lotes.Count() == 0 && animales.Count() == 0)
+                {
+                    uowService.DepartamentoRepository.Delete(mapper.Map<Departamento>(entity));
+                    uowService.Save();
+                }
+                else
+                {
+                    throw new ExceptionCore("Entidad en uso");
+                }
+
             }
             catch (ExceptionData)
             {
