+++ back/Subasta.core/services/MunicipioService.cs
@@ -13,15 +13,24 @@ using System.Text;
 
 namespace Subasta.core.services
 {
-    public class MunicipioService: IMunicipioService
+    public class MunicipioService : IMunicipioService
     {
         readonly IMapper mapper;
         readonly IUnitOfWork uowService;
-
-        public MunicipioService(IMapper mapper, IUnitOfWork uowService)
+        readonly IClienteRepository clienteRepository;
+        readonly IEventoRepository eventoRepository;
+        readonly ILoteRepository loteRepository;
+        readonly IAnimalRepository animalRepository;
+        public MunicipioService(IMapper mapper, IUnitOfWork uowService,
+             IClienteRepository clienteRepository, IEventoRepository eventoRepository,
+            ILoteRepository loteRepository, IAnimalRepository animalRepository)
         {
             this.mapper = mapper;
             this.uowService = uowService;
+            this.clienteRepository = clienteRepository;
+            this.eventoRepository = eventoRepository;
+            this.loteRepository = loteRepository;
+            this.animalRepository = animalRepository;
         }
 
         public void Add(MunicipioDto dto)
@@ -42,12 +51,41 @@ namespace Subasta.core.services
             }
         }
 
+        public int AddWithReturn(MunicipioDto dto)
+        {
+            try
+            {
+                return uowService.MunicipioRepository.Add(mapper.Map<Municipio>(dto));
+            }
+            catch (ExceptionData)
+            {
+                throw;
+            }
+            catch (Exception ex)
+            {
+
+                throw new ExceptionCore("error al intentar agregar el municipio", ex);
+            }
+        }
+
         public void Delete(MunicipioDto entity)
         {
             try
             {
-                uowService.MunicipioRepository.Delete(mapper.Map<Municipio>(entity));
-                uowService.Save();
+                var clientes = clienteRepository.GetAll().Where(c => c.MunicipioId == entity.DepartamentoId);
+                var eventos = eventoRepository.GetAll().Where(d => d.MunicipioId == entity.DepartamentoId);
+                var lotes = loteRepository.GetAll().Where(d => d.MunicipioId == entity.DepartamentoId);
+                var animales = animalRepository.GetAll().Where(d => d.MunicipioId == entity.DepartamentoId);
+
+                if (clientes.Count() == 0 && eventos.Count() == 0 && lotes.Count() == 0 && animales.Count() == 0)
+                {
+                    uowService.MunicipioRepository.Delete(mapper.Map<Municipio>(entity));
+                    uowService.Save();
+                }
+                else
+                {
+                    throw new ExceptionCore("Entidad en uso");
+                }
             }
             catch (ExceptionData)
             {
